requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.25
python-dotenv==1.0.1